refactored data.lua
